move some proxy-related tests from HttpClientHandlerTest.cs to HttpClientHandlerTest.Proxy.cs
