+++ src/libraries/Common/tests/System/Net/Http/HttpClientHandlerTest.cs
@@ -252,166 +252,6 @@ namespace System.Net.Http.Functional.Tests
             Assert.True(connectionAccepted);
         }
 
-        [Theory]
-        [InlineData("1.2.3.4")]
-        [InlineData("1.2.3.4:8080")]
-        [InlineData("[::1234]")]
-        [InlineData("[::1234]:8080")]
-        public async Task ProxiedIPAddressRequest_NotDefaultPort_CorrectlyFormatted(string host)
-        {
-            string uri = "http://" + host;
-            bool connectionAccepted = false;
-
-            if (UseVersion == HttpVersion30)
-            {
-                return;
-            }
-
-            await LoopbackServer.CreateClientAndServerAsync(async proxyUri =>
-            {
-                using (HttpClientHandler handler = CreateHttpClientHandler())
-                using (HttpClient client = CreateHttpClient(handler))
-                {
-                    handler.Proxy = new WebProxy(proxyUri);
-                    try { await client.GetAsync(uri); } catch { }
-                }
-            }, server => server.AcceptConnectionAsync(async connection =>
-            {
-                connectionAccepted = true;
-                List<string> headers = await connection.ReadRequestHeaderAndSendResponseAsync();
-                Assert.Contains($"GET {uri}/ HTTP/1.1", headers);
-            }));
-
-            Assert.True(connectionAccepted);
-        }
-
-        public static IEnumerable<object[]> DestinationHost_MemberData()
-        {
-            yield return new object[] { "nosuchhost.invalid" };
-            yield return new object[] { "1.2.3.4" };
-            yield return new object[] { "[::1234]" };
-        }
-
-        [Theory]
-        [MemberData(nameof(DestinationHost_MemberData))]
-        public async Task ProxiedRequest_DefaultPort_PortStrippedOffInUri(string host)
-        {
-            string addressUri = $"http://{host}:80/";
-            string expectedAddressUri = $"http://{host}/";
-            bool connectionAccepted = false;
-
-            if (UseVersion == HttpVersion30)
-            {
-                return;
-            }
-
-            await LoopbackServer.CreateClientAndServerAsync(async proxyUri =>
-            {
-                using (HttpClientHandler handler = CreateHttpClientHandler())
-                using (HttpClient client = CreateHttpClient(handler))
-                {
-                    handler.Proxy = new WebProxy(proxyUri);
-                    try { await client.GetAsync(addressUri); } catch { }
-                }
-            }, server => server.AcceptConnectionAsync(async connection =>
-            {
-                connectionAccepted = true;
-                List<string> headers = await connection.ReadRequestHeaderAndSendResponseAsync();
-                Assert.Contains($"GET {expectedAddressUri} HTTP/1.1", headers);
-            }));
-
-            Assert.True(connectionAccepted);
-        }
-
-        [Theory]
-        [MemberData(nameof(DestinationHost_MemberData))]
-        public async Task ProxyTunnelRequest_PortSpecified_NotStrippedOffInUri(string host)
-        {
-            // Https proxy request will use CONNECT tunnel, even the default 443 port is specified, it will not be stripped off.
-            string requestTarget = $"{host}:443";
-            string addressUri = $"https://{host}/";
-            bool connectionAccepted = false;
-
-            if (UseVersion == HttpVersion30)
-            {
-                return;
-            }
-
-            await LoopbackServer.CreateClientAndServerAsync(async proxyUri =>
-            {
-                using (HttpClientHandler handler = CreateHttpClientHandler())
-                using (HttpClient client = CreateHttpClient(handler))
-                {
-                    handler.Proxy = new WebProxy(proxyUri);
-                    handler.ServerCertificateCustomValidationCallback = TestHelper.AllowAllCertificates;
-                    try { await client.GetAsync(addressUri); } catch { }
-                }
-            }, server => server.AcceptConnectionAsync(async connection =>
-            {
-                connectionAccepted = true;
-                List<string> headers = await connection.ReadRequestHeaderAndSendResponseAsync();
-                Assert.Contains($"CONNECT {requestTarget} HTTP/1.1", headers);
-            }));
-
-            Assert.True(connectionAccepted);
-        }
-
-        [Theory]
-        [InlineData(true)]
-        [InlineData(false)]
-        public async Task ProxyTunnelRequest_UserAgentHeaderAdded(bool addUserAgentHeader)
-        {
-            if (IsWinHttpHandler)
-            {
-                return; // Skip test since the fix is only in SocketsHttpHandler.
-            }
-
-            if (UseVersion == HttpVersion30)
-            {
-                return;
-            }
-
-            string host = "nosuchhost.invalid";
-            string addressUri = $"https://{host}/";
-            bool connectionAccepted = false;
-
-            await LoopbackServer.CreateClientAndServerAsync(async proxyUri =>
-            {
-                using (HttpClientHandler handler = CreateHttpClientHandler())
-                using (var client = new HttpClient(handler))
-                {
-                    handler.Proxy = new WebProxy(proxyUri);
-                    handler.ServerCertificateCustomValidationCallback = TestHelper.AllowAllCertificates;
-                    if (addUserAgentHeader)
-                    {
-                        client.DefaultRequestHeaders.UserAgent.Add(new ProductInfoHeaderValue("Mozilla", "5.0"));
-                    }
-                    try
-                    {
-                        await client.GetAsync(addressUri);
-                    }
-                    catch
-                    {
-                    }
-                }
-            }, server => server.AcceptConnectionAsync(async connection =>
-            {
-                connectionAccepted = true;
-                List<string> headers = await connection.ReadRequestHeaderAndSendResponseAsync();
-                Assert.Contains($"CONNECT {host}:443 HTTP/1.1", headers);
-                if (addUserAgentHeader)
-                {
-                    Assert.Contains("User-Agent: Mozilla/5.0", headers);
-                }
-                else
-                {
-                    Assert.DoesNotContain("User-Agent:", headers);
-                }
-            }));
-
-            Assert.True(connectionAccepted);
-        }
-
         public static IEnumerable<object[]> SecureAndNonSecure_IPBasedUri_MemberData() =>
             from address in new[] { IPAddress.Loopback, IPAddress.IPv6Loopback }
             from useSsl in BoolValues
